% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes.R
\name{create_kernel}
\alias{create_kernel}
\title{Create U-statistic kernel}
\usage{
create_kernel(kernel_function, kernel_name, symmetric = TRUE)
}
\arguments{
\item{kernel_function}{Function taking two arguments}

\item{kernel_name}{Character name for the kernel}

\item{symmetric}{Logical indicating if kernel is symmetric}
}
\value{
UStatKernel object
}
\description{
Create U-statistic kernel
}
