% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes.R
\docType{class}
\name{DecoupleResult-class}
\alias{DecoupleResult-class}
\title{DecoupleResult Class}
\description{
S4 class storing results from decoupling analysis
}
\section{Slots}{

\describe{
\item{\code{original_stat}}{Numeric value of the original U-statistic}

\item{\code{decoupled_distribution}}{Numeric vector of decoupled statistics}

\item{\code{kernel_name}}{Character string describing the kernel used}

\item{\code{method}}{Character string describing the decoupling method}

\item{\code{p_value}}{Numeric p-value from hypothesis testing}
}}

