% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupling_algorithms.R
\name{hamming_distance_kernel}
\alias{hamming_distance_kernel}
\title{Hamming Distance Kernel}
\usage{
hamming_distance_kernel(seq1, seq2)
}
\arguments{
\item{seq1}{First DNA sequence (character vector)}

\item{seq2}{Second DNA sequence (character vector)}
}
\value{
Numeric Hamming distance
}
\description{
Kernel function for calculating Hamming distance between DNA sequences
}
