% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_example_sequences}
\alias{load_example_sequences}
\title{Load Example DNA Sequences}
\usage{
load_example_sequences()
}
\value{
List of DNA sequences
}
\description{
Load example DNA sequences from package data
}
