% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_example.R
\name{plot,DecoupleResult,ANY-method}
\alias{plot,DecoupleResult,ANY-method}
\title{Plot DecoupleResult}
\usage{
\S4method{plot}{DecoupleResult,ANY}(x, main = "Decoupling Analysis", ...)
}
\arguments{
\item{x}{DecoupleResult object}

\item{main}{Plot title}

\item{...}{Additional arguments to plot}
}
\description{
Plot DecoupleResult
}
