% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes.R
\name{show,DecoupleResult-method}
\alias{show,DecoupleResult-method}
\title{Show method for DecoupleResult}
\usage{
\S4method{show}{DecoupleResult}(object)
}
\arguments{
\item{object}{DecoupleResult object to display}
}
\description{
Show method for DecoupleResult
}
