% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes.R
\docType{class}
\name{UStatKernel-class}
\alias{UStatKernel-class}
\title{UStatKernel Class}
\description{
S4 class for U-statistic kernel functions
}
\section{Slots}{

\describe{
\item{\code{kernel_function}}{Function implementing the kernel}

\item{\code{kernel_name}}{Character string name of the kernel}

\item{\code{symmetric}}{Logical indicating if kernel is symmetric}
}}

