% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Validate Input Data}
\usage{
validate_input_data(x, kernel)
}
\arguments{
\item{x}{Input data to validate}

\item{kernel}{Kernel to validate against}
}
\description{
Validate input data for decoupling analysis
}
